% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_graph_of_winner_pct_margins.R
\name{bargraph_of_margins}
\alias{bargraph_of_margins}
\title{Create bar graphs of winner's margin per district by percent or raw votes}
\usage{
bargraph_of_margins(
  election_df,
  district_col,
  margin_col,
  headline = "",
  palette = c("#9970ab", "#5aae61", "#f7f7f7"),
  brewer_palette = FALSE,
  winner_col = Winner,
  theme = theme_minimal(),
  order_bars_desc = FALSE
)
}
\arguments{
\item{election_df}{data frame of election results generated by wrangle_results() or wrangle_more_cols()}

\item{district_col}{name of the election district column NOT in quotation marks}

\item{margin_col}{name of the column with percent or vote margins NOT in quotation marks}

\item{headline}{character string for desired graph headline, defaults to ""}

\item{palette}{if brewer_palette is FALSE, character string of hex colors for winner, loser, and ties
if brewer palette is TRUE, character string with RColorBrewer palette name or integer with palette number.}

\item{brewer_palette}{logical if TRUE use an RColorBrewer palette}

\item{winner_col}{name of column with the winner for each district NOT in quotation marks, defaults to Winner.}

\item{theme}{ggplot theme function. Defaults to theme_minimal() but allows for easy change of theme while still using function.}

\item{order_bars_desc}{logical should bars be ordered by descending value? Defaults to FALSE}
}
\value{
ggplot object
}
\description{
This function expects data in the format generated by the wrangle_results() or wrangle_more_cols() functions.
}
\examples{
my_file <- system.file("extdata", "FakeElectionResults.xlsx", package = "elections2")
my_election_results <- wrangle_results(my_file)
bargraph_of_margins(my_election_results, Precinct, Yes_Pct_Margin, order_bars_desc = TRUE, headline = "Yes Margins by Precinct")
}
