% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{framingham_gis}
\alias{framingham_gis}
\title{Framingham precinct sf object}
\format{
An sf object with 18 rows of 7 variables:
\describe{
  \item{district}{number with district number}
  \item{objectid}{number with city object ID}
  \item{precinct}{character string P and precinct number}
  \item{precncts_i}{number}
  \item{shape_star}{number}
  \item{shape_stle}{number}
  \item{geometry}{sfc_POLYGON with precinct geospatial data}
}
}
\source{
\url{https://data.framinghamma.gov/Community-Development/Framingham-Districts-and-Precincts/9pzx-4i9g}
}
\usage{
framingham_gis
}
\description{
A geospatial sf object from a shapefile via the city of Framingham data portal with precinct geometry
}
\keyword{datasets}
