% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_election_results.R
\name{map_election_results_2layers}
\alias{map_election_results_2layers}
\title{Generate map of election results with 1 candidate per layer}
\usage{
map_election_results_2layers(
  joined_df,
  win_col,
  lose_col,
  election_district_col,
  the_winner_palette = "Purples",
  the_loser_palette = "Greens",
  map_tiles = "CartoDB.Positron"
)
}
\arguments{
\item{joined_df}{sf object shapefile of election districts that's been joined with election result data.table generated by wrangle_results() function}

\item{win_col}{character string name of column with winner's data by Pct or Margin}

\item{lose_col}{character string name of column with loser's data by Pct or Margin}

\item{election_district_col}{character string name of column with election district data - should be name of 1st column from the wrangle_results() data frame}

\item{the_winner_palette}{character string of colors for leaflet colorNumeric function}

\item{the_loser_palette}{character string of colors for leaflet colorNumeric function}

\item{map_tiles}{character string for available leaflet tile providers, defaults to "CartoDB.Positron". See http://leaflet-extras.github.io/leaflet-providers/preview/ for options.}
}
\value{
leaflet map object
}
\description{
Requires data from wrangle_results() function joined with a shapefile. See maps vignette for more info.
}
