% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytest.R, R/wrangle_more_cols.R
\name{mytest}
\alias{mytest}
\title{data.table test function}
\usage{
mytest(x)

mytest(x)
}
\arguments{
\item{x}{string}

\item{mycols}{string}

\item{mydata}{data.table}
}
\value{
data.table

string
}
\description{
data.table test function

Can I add a new function
}
