% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{na2zero2}
\alias{na2zero2}
\title{Convert all NA values to z}
\usage{
na2zero2(mydf)
}
\arguments{
\item{mydf}{data frame}
}
\value{
data frame with NAs converted to 0s
}
\description{
Convert all NA values to z
}
