% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table_more_cols.R
\name{quick_table_more_cols}
\alias{quick_table_more_cols}
\title{Generate table of results with more than 2 columns, including sparkline bar graphs for percent data}
\usage{
quick_table_more_cols(
  election_df,
  graph_cols,
  show_pcts = FALSE,
  pagination = 20,
  show_sparklines = TRUE,
  use_regex_searching = TRUE
)
}
\arguments{
\item{election_df}{data table of election results generated by wrangle_more_cols() function}

\item{graph_cols}{vector of character strings with names of columns containing vote or percent results}

\item{show_pcts}{logical whether election_df data is showing percents as opposed to vote totals. Defaults to FALSE.}

\item{pagination}{integer number of rows to display per page, defaults to 20}

\item{show_sparklines}{logical if you want table to include a column with mini bar charts. Defaults to TRUE.}
}
\value{

}
\description{
Generate table of results with more than 2 columns, including sparkline bar graphs for percent data
}
\examples{
myfile <- system.file("extdata", "Fake3Columns.xlsx", package = "elections2")
mydata <- wrangle_more_cols(myfile, c("Red", "Blue", "Orange", "Green", "Purple"), show_pcts = TRUE)
quick_table_more_cols(mydata, c("Red", "Blue", "Orange", "Green", "Purple"), show_pcts = TRUE)
}
