% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table.R
\name{quick_table_w_bargraphs}
\alias{quick_table_w_bargraphs}
\title{Create interactive table of results with bar graph for win pct margin}
\usage{
quick_table_w_bargraphs(
  election_df,
  pagination = 20,
  win_color = "#c2a5cf",
  lose_color = "#a6dba0"
)
}
\arguments{
\item{election_df}{data frame with results from wrangle_results() or wrangle_more_cols()}

\item{pagination}{integer number of results on page, defaults to 20}

\item{win_color}{character string with color hex code}

\item{lose_color}{character string with color hex code}
}
\value{
reactable interactive HTML table
}
\description{
Create interactive table of results with bar graph for win pct margin
}
