% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_more_cols.R
\name{wrangle_more_cols}
\alias{wrangle_more_cols}
\title{Get winner for race with three or more candidates}
\usage{
wrangle_more_cols(
  election_results_file,
  votes_cols,
  show_pcts = FALSE,
  show_runnerup = TRUE,
  show_margin = TRUE,
  show_margin_2vs3 = FALSE,
  use_column_numbers = FALSE
)
}
\arguments{
\item{election_results_file}{character string with name of Excel or CSV file in format with election district in 1st col. You can also use the name of an R data frame if you already have your results data in R, but the object name MUST be in quotation marks.}

\item{votes_cols}{vector of character strings with names of columns containing votes (or numeric vector if use_column_numbers is TRUE)}

\item{show_pcts}{logical TRUE to display results as percents. Defaults to FALSE for vote totals.}

\item{show_runnerup}{logical TRUE to show 2nd place finisher. Defaults to TRUE.}

\item{show_margin}{logical TRUE to show margin between 1st and 2nd place. Defaults to TRUE.}

\item{show_margin_2vs3}{logical TRUE to show margin between 2nd and 3rd place. Defaults to FALSE}

\item{use_column_numbers}{logical TRUE to use column index numbers instead of character strings with column names in votes_cols. Defaults to FALSE.}
}
\value{
data table with results, totals, and district winner by number of votes
}
\description{
Get winner for race with three or more candidates
}
\examples{
myfile <- system.file("extdata", "Fake3Columns.xlsx", package = "elections2")
mydata <- wrangle_more_cols(myfile, c("Red", "Blue", "Green"), show_pcts = TRUE, show_runnerup = FALSE)
}
