% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_more_cols.R
\name{wrangle_more_cols_pcts}
\alias{wrangle_more_cols_pcts}
\title{Get winner and percents by district for race with three or more candidates or options}
\usage{
wrangle_more_cols_pcts(election_results_file, start_col, end_col)
}
\arguments{
\item{election_results_file}{character string with name of Excel or CSV file. File must be in format with election district in column one, and candidate or option raw vote results in columns adjacent to each other.}

\item{start_col}{character string name of column starting the results}

\item{end_col}{character string name of column ending the results}
}
\value{
data frame with results, totals, and district winner by percents
}
\description{
Get winner and percents by district for race with three or more candidates or options
}
