% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mytest.R
\name{mytest}
\alias{mytest}
\title{data.table test function}
\usage{
mytest(mycols, mydata)
}
\arguments{
\item{mycols}{string}

\item{mydata}{data.table}
}
\value{
data.table
}
\description{
data.table test function
}
