% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table.R
\name{quick_table}
\alias{quick_table}
\title{Create a quick searchable, sortable table of results with the DT package}
\usage{
quick_table(
  election_df,
  pagination = 20,
  win_color = "#c2a5cf",
  lose_color = "#a6dba0",
  use_regex_searching = TRUE,
  n_turnout_cols = 0
)
}
\arguments{
\item{election_df}{data frame of election results generated by wrangle_results() including Total row}

\item{pagination}{integer how many rows to display on each color}

\item{win_color}{character string for color bar in winner's percent column, defaults to "#c2a5cf" (a light purple). Use "none" for no colored bar.}

\item{lose_color}{character string for color bar in winner's percent column, defaults to "#a6dba0" (a light green). Use "none" for no colored bar.}

\item{use_regex_searching}{logical whether to allow searching by regular expression, defaults to TRUE}

\item{n_turnout_cols}{integer how many turnout columns have been joined to election_df: 0 for no data about turnout; 1 for a column with turnout percent only; or 3 for columns containing total votes cast, total registered voters, and percent turnout. Defaults to 0.}
}
\value{
html widget with interactive table of results
}
\description{
Features colored bars in percent columns and allow for searching by regular expression
}
\examples{
my_election_results <- wrangle_results(system.file("extdata", "FakeElectionResults.xlsx", package = "elections2"))
quick_table(my_election_results)
quick_table(my_election_results, win_color = "none", lose_color = "none")
}
