% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quick_table_more_cols.R
\name{quick_table_more_cols}
\alias{quick_table_more_cols}
\title{Generate table of results with more than 2 columns, including sparkline bar graphs for percent data}
\usage{
quick_table_more_cols(
  election_df,
  graph_cols,
  num_format = "pct",
  pagination = 20
)
}
\arguments{
\item{election_df}{data table of election results generated by wrangle_more_cols() function}

\item{graph_cols}{vector of character strings with names of columns containing vote or percent results}

\item{num_format}{character string either "pct" or "votes" - whether election_df contains vote totals or percents}

\item{pagination}{integer number of rows to display per page, defaults to 20}
}
\value{

}
\description{
Generate table of results with more than 2 columns, including sparkline bar graphs for percent data
}
\examples{
myfile <- system.file("extdata", "Fake3Columns.xlsx", package = "elections2")
mydt <- wrangle_more_cols(myfile, c("Red", "Blue", "Orange", "Green", "Purple"))
quick_table_more_cols(mydt, c("Red", "Blue", "Orange", "Green", "Purple"))
}
