% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle_results.R
\name{wrangle_results}
\alias{wrangle_results}
\title{Get election winner and loser by election district from spreadsheet or CSV file}
\usage{
wrangle_results(election_results_file)
}
\arguments{
\item{election_results_file}{Excel or CSV file of election results with only two candidates (or yes/no for ballot questions). Should have only 3 columns: column 1 the precinct, city, county, state, or other election district; and columns 2 and 3 the candidate names (or yes and no, or any other two choices). Do NOT include any column or row totals.}
}
\value{
data.table with additional columns for total votes, winner, winner percent, loser's percent, winner's vote margin, and winner's percentage point margin.
}
\description{
Get election winner and loser by election district from spreadsheet or CSV file
}
